requirements.txt
absl-py==0.12.0
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
backtrader==1.9.76.123
base58==2.1.0
beautifulsoup4==4.9.3
bleach==3.3.0
blinker==1.4
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
clang==5.0
click==7.1.2
cloudpickle==1.6.0
cmdstanpy==0.9.5
contextvars==2.4
convertdate==2.3.2
cssselect==1.1.0
cufflinks==0.17.3
cycler==0.10.0
Cython==0.29.22
dask==2021.3.0
dask[distributed]
decorator==5.0.3
defusedxml==0.7.1
empyrical==0.5.5
entrypoints==0.3
ephem==3.7.7.1
fake-useragent==0.1.11
finviz==1.4.2
finvizfinance==0.9.10
feedparser==6.0.2
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
HeapDict==1.0.1
hijri-converter==2.1.1
holidays==0.11.1
html5lib==1.1
idna==2.10
immutables==0.15
importlib-metadata==3.10.0
inflection==0.5.1
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
more-itertools==8.7.0
mplfinance==0.12.7a12
msgpack==1.0.2
multitasking==0.0.9
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk
notebook==6.4.3
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.3.3
pandas-datareader==0.10.0
pandocfilters==1.4.3
parse==1.19.0
parso==0.7.0
pathlib==1.0.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
plotly==5.3.1
pmdarima==1.8.1
prometheus-client==0.10.0
prompt-toolkit==3.0.18
fbprophet==0.7.1
protobuf==3.15.7
psutil==5.8.0
# ptyprocess==0.7.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.6.1
pyee==8.1.0
pyfolio==0.9.2
pyforest==1.0.3
Pygments==2.8.1
PyMeeus==0.5.11
pyparsing==2.4.7
pyppeteer==0.2.6
pyqt5==5.12.0
pyqtwebengine==5.12.0
pyquery==1.4.3
pyrsistent==0.17.3
pystan==2.19.1.1
python-dateutil==2.8.1
pytrends==4.7.3
pytz==2021.1
# PyYAML==5.4.1
pyzmq==22.0.3
Quandl==3.6.1
requests==2.25.1
requests-futures==1.0.0
requests-html==0.10.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.5.0
setuptools-git==1.2
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
smmap==4.0.0
sortedcontainers==2.3.0
soupsieve==2.2.1
statsmodels==0.12.2
streamlit==0.89.0
tabulate==0.8.9
tblib==1.7.0
tensorboard==2.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
# terminado==0.12.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.59.0
traitlets==4.3.3
tscv==0.1.1
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.6
validators==0.18.2
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==9.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.3.3
yahoo-fin==0.8.9.1
yahooquery==2.2.15
yfinance==0.1.63
